[#3088] Move Selenium-based test to Playwright; remove old functional suite
